blackjack/requirements.txt
django~=5.1.4
psycopg2
djangorestframework-3.15.2

djangorestframework~=3.15.2